moved pulseoperations from calendarfragment into seperate class and improved imports on all classes
