+++ app/src/main/java/io/punchtime/punchtime/logic/loaders/TimerLoader.java
@@ -6,10 +6,6 @@ import android.content.Intent;
 import android.content.IntentFilter;
 import android.support.v4.content.Loader;
 
-import java.text.SimpleDateFormat;
-import java.util.Date;
-import java.util.Locale;
-
 /**
  * Created by elias on 28/04/16.
  * for project: Punchtime
