+++ app/src/main/java/io/punchtime/punchtime/ui/activities/PermissionErrorActivity.java
@@ -13,7 +13,6 @@ import android.support.v4.content.ContextCompat;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.TextView;
-import com.google.android.gms.common.api.GoogleApiClient;
 
 import io.punchtime.punchtime.R;
