+++ app/src/main/java/io/punchtime/punchtime/ui/fragments/WeekFragment.java
@@ -34,6 +34,7 @@ public class WeekFragment extends CalendarFragment {
 
         setupDateTimeInterpreter(true);
 
+
         //return view
 
         return v;
